requirements.txt

streamlit==1.34.0
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain==0.1.14
langchain-google-genai==0.0.8
langchain-huggingface==0.0.6
faiss-cpu==1.8.0
google-generativeai==0.0.5
